Add `pre-commit` hooks
This commit adds the ability to run `pre-commit` on code committed to
the repo. I have run it on the repo, which has required the following
files to be updated.

- CONTRIBUTING.md
- README.md
- arviz_dashboard/__init__.py
- arviz_dashboard/elpd.py
- examples/elpd_dashboard.ipynb
- examples/posterior-marginal1d-examples.ipynb

There are several configurations we can use for `pre-commit` that can be
stored in a `pyproject.toml` file. This commit also updates the
`setup.py` file so we can install locally using the `pyproject.toml`
file instead of the legacy `setup.py` file.