pyproject (#18)
This commit adds the following new files.

- `LICENSE.txt`
- `pyproject.toml`

It removes the older style installation files below in favor of the
`pyproject.toml` file.

- `requirements.txt`
- `setup.py`

The `README.md` file has been updated to reflect how to install a
development copy of `arviz_dashboard`. We are also moving all files in
the repo so that we have a `src-layout` repository for future directions
with `tests` as well as the possibility of needing to create a monorepo
for JavaScript dependencies.